refactor(config): split config.rs into domain submodules
